requirements.txt
discord.py
Faker
peewee